requirements.txt
arabic-reshaper==3.0.0
asgiref==3.11.1
asn1crypto==1.5.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
click==8.1.7
colorthief==0.2.1
colorama==0.4.6
contourpy==1.0.7
cryptography==41.0.3
cssselect2==0.7.0
cycler==0.11.0
Django==4.1.13
django-baton==2.6.2
django-cors-headers
django-environ==0.9.0
django-filter
django-restframework==0.0.1
django-rosetta==0.9.8
django-social-share==2.3.0
django-suit==0.2.7
django-tinymce==3.6.1
djangorestframework==3.14.0
fontawesome-free==5.15.4
fontawesomefree==6.2.1
fonttools==4.38.0
html2pdf==0.1
html5lib==1.1
idna==3.4
kiwisolver==1.4.4
lxml==4.9.3
matplotlib==3.6.3
PyMySQL==1.1.1
numpy==1.24.2
opencv-python==4.7.0.72
oscrypto==1.3.0
packaging==23.0
phonenumbers==9.0.25
Pillow==9.4.0
polib==1.1.1
pycparser==2.21
pyHanko==0.20.0
pyhanko-certvalidator==0.23.0
pypdf==3.15.4
pypng==0.20220715.0
pyparsing==3.0.9
python-bidi==0.4.2
python-dateutil==2.8.2
python-gettext==4.1
pytz==2022.7.1
PyYAML==6.0.1
qrcode==7.4.2
reportlab==3.6.13
requests==2.28.2
rosetta==0.3
six==1.16.0
sqlparse==0.4.3
svglib==1.5.1
tinycss2==1.2.1
transliterate==1.10.2
typing_extensions==4.4.0
tzdata==2023.3
tzlocal==5.0.1
uritools==4.0.2
urllib3==1.26.14
webencodings==0.5.1
xhtml2pdf==0.2.11
daphne
channels
channels-redis
django-redis
kafka-python
redis
python-dotenv
psycopg2-binary
Pillow
flake8
whitenoise
celery
flower